requirements/core.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/core.in
#
aiofiles==23.1.0
    # via
    #   -r requirements/core.in
    #   sanic
aiosqlite==0.17.0
    # via tortoise-orm
asyncpg==0.27.0
    # via tortoise-orm
azure-core==1.26.3
    # via azure-storage-blob
azure-storage-blob==12.15.0
    # via -r requirements/core.in
bcrypt==4.0.1
    # via -r requirements/core.in
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
cryptography==40.0.1
    # via azure-storage-blob
dnspython==2.3.0
    # via email-validator
email-validator==2.0.0
    # via pydantic
httptools==0.5.0
    # via sanic
idna==3.4
    # via
    #   email-validator
    #   requests
iso8601==1.1.0
    # via tortoise-orm
isodate==0.6.1
    # via azure-storage-blob
joblib==1.2.0
    # via scikit-learn
multidict==6.0.4
    # via sanic
numpy==1.24.2
    # via
    #   scikit-learn
    #   scipy
pycparser==2.21
    # via cffi
pydantic[email]==1.10.7
    # via -r requirements/core.in
pyjwt==2.6.0
    # via -r requirements/core.in
pypika-tortoise==0.1.6
    # via tortoise-orm
pytz==2023.3
    # via tortoise-orm
pyyaml==6.0
    # via sanic-ext
regex==2022.10.31
    # via sentence-splitter
requests==2.28.2
    # via azure-core
sanic[ext]==22.12.0
    # via -r requirements/core.in
sanic-ext==23.3.0
    # via sanic
sanic-routing==22.8.0
    # via sanic
scikit-learn==1.2.1
    # via -r requirements/core.in
scipy==1.10.1
    # via scikit-learn
sentence-splitter==1.4
    # via -r requirements/core.in
six==1.16.0
    # via
    #   azure-core
    #   isodate
threadpoolctl==3.1.0
    # via scikit-learn
tortoise-orm[asyncpg]==0.19.3
    # via -r requirements/core.in
typing-extensions==4.5.0
    # via
    #   aiosqlite
    #   azure-core
    #   azure-storage-blob
    #   pydantic
ujson==5.7.0
    # via sanic
urllib3==1.26.14
    # via requests
uvloop==0.17.0
    # via sanic
websockets==10.4
    # via sanic
